requirements.txt
bokeh==3.1.0
h5py==3.7.0
matplotlib==3.7.4
netcdf4==1.6.3
numpy==1.21.5
openpyxl==3.0.9
pandas==1.4.2
psutil==5.9.0
pyparsing==3.0.9
PyYAML==6.0
ray==2.1.0
scikit_learn==1.2.2
scipy==1.10.1
torch==1.13.1
torchvision==0.15.2
torchmetrics==0.8.1
xarray==0.20.1
h5netcdf==1.1.0

